+++ platformio/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (C) Ivan Kravets <me@ikravets.com>
 # See LICENSE for details.
 
-VERSION = (2, 2, "0.dev3")
+VERSION = (2, 2, "0.dev4")
 __version__ = ".".join([str(s) for s in VERSION])
 
 __title__ = "platformio"